refactor: move FX status helpers to shared utils
