+++ frontend/src/budgets/components/shared/BudgetFxStatusTooltip.tsx
@@ -2,7 +2,7 @@ import type { SyntheticEvent } from 'react'
 import type { FxStatus } from '@/api/shared/fx'
 import IconTooltip from '@/components/IconTooltip'
 import { getBudgetUtilizationFxStatusMessage } from '@/budgets/utils/fxTooltipMessages'
-import { formatMissingFxPairs, getFxStatusTone } from '@/dashboard/utils/fxStatus'
+import { formatMissingFxPairs, getFxStatusTone } from '@/utils/fxStatus'
 
 export default function BudgetFxStatusTooltip({
   fxStatus,
