+++ frontend/src/insights/components/FxStatusBadge.tsx
@@ -1,6 +1,6 @@
 import type { FxStatus } from '@/api/shared/fx'
 import IconTooltip from '@/components/IconTooltip'
-import { formatMissingFxPairs, getFxStatusMessage, getFxStatusTone } from '@/dashboard/utils/fxStatus'
+import { formatMissingFxPairs, getFxStatusMessage, getFxStatusTone } from '@/utils/fxStatus'
 
 type FxStatusBadgeProps = {
   label: string
