+++ frontend/src/transactions/components/topBand/NetFlowSummary.tsx
@@ -2,7 +2,7 @@ import { useLayoutEffect, useRef, useState } from 'react'
 import { ArrowDownLeft, ArrowUpRight } from 'lucide-react'
 import type { FxStatus } from '@/api/shared/fx'
 import IconTooltip from '@/components/IconTooltip'
-import { formatMissingFxPairs, getFxStatusTone } from '@/dashboard/utils/fxStatus'
+import { formatMissingFxPairs, getFxStatusTone } from '@/utils/fxStatus'
 import { getCashFlowFxStatusMessage } from '@/transactions/utils/fxTooltipMessages'
 import { formatCurrency } from '@/utils/formatCurrency'
