+++ frontend/src/transactions/components/topBand/TopCategoriesChart.tsx
@@ -22,7 +22,7 @@ import {
   type DeferredChartTooltipOverlayHandle,
 } from '@/components/charts/DeferredChartTooltipOverlay'
 import IconTooltip from '@/components/IconTooltip'
-import { formatMissingFxPairs, getFxStatusTone } from '@/dashboard/utils/fxStatus'
+import { formatMissingFxPairs, getFxStatusTone } from '@/utils/fxStatus'
 import { formatCurrency } from '@/utils/formatCurrency'
 import {
   TOP_CATEGORY_AXIS_AVG_CHAR_WIDTH,
